+++ src/ocaml/typing/typetexp.ml
@@ -188,8 +188,20 @@ let instance_poly_univars env loc vars =
 type policy = Fixed | Extensible | Univars
 
 let rec transl_type env policy styp =
-  Builtin_attributes.warning_scope styp.ptyp_attributes
-    (fun () -> transl_type_aux env policy styp)
+  Msupport.with_saved_types
+    ~warning_attribute:styp.ptyp_attributes ?save_part:None
+    (fun () -> 
+       try
+         transl_type_aux env policy styp
+       with exn ->
+         Msupport.raise_error exn;
+           { ctyp_desc = Ttyp_any;
+             ctyp_type = new_global_var ();
+             ctyp_env = env;
+             ctyp_loc = styp.ptyp_loc;
+             ctyp_attributes = [];
+           }
+    )
 
 and transl_type_aux env policy styp =
   let loc = styp.ptyp_loc in