+++ tests/test-dirs/issue1109.t/run.t
@@ -1,5 +1,3 @@
-FIXME: expected 'a -> 'a
-
   $ $MERLIN single type-enclosing -position 5:11 -verbosity 0 \
   > -filename ./issue1109.ml < ./issue1109.ml | jq ".value[0:2]"
   [
@@ -12,7 +10,7 @@ FIXME: expected 'a -> 'a
         "line": 5,
         "col": 14
       },
-      "type": "'a -> 'b",
+      "type": "'a -> 'a",
       "tail": "no"
     },
     {